+++ src/main/java/Kraisie/SceneController/AddSeriesController.java
@@ -97,7 +97,7 @@ public class AddSeriesController extends Controller {
 	@FXML
 	private void back() {
 		try {
-			openScene((Stage) backButton.getScene().getWindow(), "/FXML/MainSeries.fxml", "Series Control Panel");
+			openMain((Stage) backButton.getScene().getWindow());
 		} catch (IOException e) {
 			popUp.showError("Failed to open the scene!", getStackTrace(e), true, (Stage) backButton.getScene().getWindow());
 		}