+++ src/main/java/Kraisie/SceneController/FinishedController.java
@@ -88,7 +88,7 @@ public class FinishedController extends Controller {
 	@FXML
 	private void back() {
 		try {
-			openScene((Stage) backButton.getScene().getWindow(), "/FXML/MainSeries.fxml", "Series Control Panel");
+			openMain((Stage) backButton.getScene().getWindow());
 		} catch (IOException e) {
 			popUp.showError("Failed to open the scene!", getStackTrace(e), true, (Stage) backButton.getScene().getWindow());
 		}