+++ src/main/java/Kraisie/SceneController/SearchController.java
@@ -348,7 +348,7 @@ public class SearchController extends Controller {
 	@FXML
 	private void back() {
 		try {
-			openScene((Stage) backButton.getScene().getWindow(), "/FXML/MainSeries.fxml", "Series Control Panel");
+			openMain((Stage) backButton.getScene().getWindow());
 		} catch (IOException e) {
 			popUp.showError("Failed to open the scene!", getStackTrace(e), true, (Stage) backButton.getScene().getWindow());
 		}