+++ src/main/java/Kraisie/SceneController/SelectController.java
@@ -285,7 +285,7 @@ public class SelectController extends Controller {
 	@FXML
 	private void backToMain() {
 		try {
-			openScene((Stage) backButton.getScene().getWindow(), "/FXML/MainSeries.fxml", "Series Control Panel");
+			openMain((Stage) backButton.getScene().getWindow());
 		} catch (IOException e) {
 			popUp.showError("Failed to open the scene!", getStackTrace(e), true, (Stage) backButton.getScene().getWindow());
 		}